+++ sync-for-expenses/CodatSyncExpenses/Models/Operations/CreateAccountRequest.cs
@@ -8,16 +8,16 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatSyncExpenses.Models.Operations
+namespace Codat.Sync.Expenses.Models.Operations
 {
-    using CodatSyncExpenses.Models.Shared;
-    using CodatSyncExpenses.Utils;
+    using Codat.Sync.Expenses.Models.Shared;
+    using Codat.Sync.Expenses.Utils;
     
     public class CreateAccountRequest
     {
 
         [SpeakeasyMetadata("request:mediaType=application/json")]
-        public Account? Account { get; set; }
+        public AccountPrototype? AccountPrototype { get; set; }
 
         /// <summary>
         /// Allow a sync upon push completion.