+++ sync-for-expenses/CodatSyncExpenses/Models/Operations/ListCustomersResponse.cs
@@ -8,9 +8,9 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatSyncExpenses.Models.Operations
+namespace Codat.Sync.Expenses.Models.Operations
 {
-    using CodatSyncExpenses.Models.Shared;
+    using Codat.Sync.Expenses.Models.Shared;
     using System.Net.Http;
     using System;
     
@@ -25,7 +25,7 @@ namespace CodatSyncExpenses.Models.Operations
         /// <summary>
         /// Success
         /// </summary>
-        public Customers? Customers { get; set; }
+        public Models.Shared.Customers? Customers { get; set; }
 
         /// <summary>
         /// Your `query` parameter was not correctly formed
@@ -40,6 +40,6 @@ namespace CodatSyncExpenses.Models.Operations
         /// <summary>
         /// Raw HTTP response; suitable for custom response parsing
         /// </summary>
-        public HttpResponseMessage? RawResponse { get; set; }
+        public HttpResponseMessage RawResponse { get; set; } = default!;
     }
 }