+++ sync-for-expenses/CodatSyncExpenses/Models/Operations/ListSyncsResponse.cs
@@ -8,20 +8,15 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatSyncExpenses.Models.Operations
+namespace Codat.Sync.Expenses.Models.Operations
 {
-    using CodatSyncExpenses.Models.Shared;
+    using Codat.Sync.Expenses.Models.Shared;
     using System.Collections.Generic;
     using System.Net.Http;
     using System;
     
     public class ListSyncsResponse
     {
-
-        /// <summary>
-        /// Success
-        /// </summary>
-        public List<CompanySyncStatus>? CompanySyncStatuses { get; set; }
 
         /// <summary>
         /// HTTP response content type for this operation
@@ -41,6 +36,11 @@ namespace CodatSyncExpenses.Models.Operations
         /// <summary>
         /// Raw HTTP response; suitable for custom response parsing
         /// </summary>
-        public HttpResponseMessage? RawResponse { get; set; }
+        public HttpResponseMessage RawResponse { get; set; } = default!;
+
+        /// <summary>
+        /// Success
+        /// </summary>
+        public List<CompanySyncStatus>? Classes { get; set; }
     }
 }