+++ sync-for-expenses/RELEASES.md
@@ -111,3 +111,13 @@ Based on:
 - [csharp v3.2.0] sync-for-expenses
 ### Releases
 - [NuGet v3.2.0] https://www.nuget.org/packages/Codat.Sync.Expenses/3.2.0 - sync-for-expenses
+
+## 2023-12-11 01:30:27
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses.yaml
+- Speakeasy CLI 1.125.2 (2.210.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v4.0.0] sync-for-expenses
+### Releases
+- [NuGet v4.0.0] https://www.nuget.org/packages/Codat.Sync.Expenses/4.0.0 - sync-for-expenses